Worker main and config
