+++ worker/.env
@@ -1,0 +1,2 @@
+PORT=5000
+MANAGER_URL=http://crackhash-manager-1:3000
